Change lots of long * to int * in CPUEMU. Now basic DPMI works for $_cpu_emu="full" on x86-64.
git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@1613 7758b857-130e-0410-ab9c-d8735a528d18

+++ src/emu-i386/simx86/codegen-x86.h
@@ -175,7 +175,7 @@ static __inline__ void PUSH(int m, void *w)
 
 // 'no-jump' version, tricky (depends on bit position)
 #define G2_4(mode, val, Cp) \
-	*((long *)(Cp))=(val); Cp+=BT24(BitDATA16, mode)
+	*((int *)(Cp))=(val); Cp+=BT24(BitDATA16, mode)
 
 
 /////////////////////////////////////////////////////////////////////////////
